Re-add templated TlbWindow class
