+++ console/tlb.cpp
@@ -3,7 +3,6 @@
 
 #include <sys/mman.h>
 #include <sys/ioctl.h>
-#include <cassert>
 #include "tlb.h"
 
 TlbHandle::TlbHandle(int fd, size_t size, const tenstorrent_noc_tlb_config &config)
@@ -52,37 +51,5 @@ TlbHandle::~TlbHandle() noexcept
 
     munmap(tlb_base, tlb_size);
     ioctl(fd, TENSTORRENT_IOCTL_FREE_TLB, &free_tlb);
-}
-
-
-TlbWindow2M::TlbWindow2M(int fd, uint16_t x, uint16_t y, uint64_t addr)
-: offset(addr & WINDOW_MASK)
-{
-    tenstorrent_noc_tlb_config config{
-        .addr = addr & ~WINDOW_MASK,
-        .x_end = x,
-        .y_end = y,
-    };
-
-    window = std::make_unique<TlbHandle>(fd, WINDOW_SIZE, config);
-}
-
-
-void TlbWindow2M::write32(uint64_t addr, uint32_t value){
-    assert(offset + addr + 4 <= WINDOW_SIZE);
-    assert(((offset + addr) % 4) == 0);
-    void *ptr = window->data() + offset + addr;
-    *reinterpret_cast<volatile uint32_t *>(ptr) = value;
-}
-
-uint32_t TlbWindow2M::read32(uint64_t addr){
-    assert(offset + addr + 4 <= WINDOW_SIZE);
-    assert(((offset + addr) % 4) == 0);
-    void *ptr = window->data() + offset + addr;
-    return *reinterpret_cast<volatile uint32_t *>(ptr);
-}
-
-uint8_t* TlbWindow2M::get_window(){
-    return window.get()->data() + offset;
 }
